requirements.txt
pytest==6.2.4
fire==0.4.0
docker_py==1.10.6
